Make tests more robust
Make sure all tests use different username when creating users (by
using a different start) so we never hit the race condition.

It was unlikely, but I just had an example where the tests failed
because it tried to create exactly the same test-alice-... user
in both prefix.feature and domain-profile.feature. So creating
uniqueness by using a timestamp down to the millisecond does not
seem to be enough when running the tests in parallel.