Add tests to verify manager as a member authorisation
I did these tests to cross verify few assumptions that were raised by
ENA team and Ania. For this, create a new feature, create a new user for
the tests to avoid a collision and add this user as a manager for the
domain, verify the behaviour of manager after and before becoming the
member of the domain.